Refactor DestinationMap to accept a list of MapPlace
The `DestinationMap` function was refactored to receive a list of `MapPlace` instead of a list of `ItineraryItem`.
A `MapPlace` data class was created to encapsulate the latitude, longitude, and name of a place.
The logic to create a `MapPlace` from either an `InterestPlace` or `ItineraryItem` was added to the companion object of the `MapPlace`.
The tab content was updated to filter and convert the places into a list of `MapPlace`.